+++ tests/system_events/test_added_to_chat.py
