+++ tests/system_events/test_chat_created.py
