+++ tests/system_events/test_deleted_from_chat.py
