+++ tests/test_base_command.py
