+++ tests/test_bot_constructing.py
