+++ tests/test_lifespan.py
